+++ tests/test_wsidicom_integration.py
@@ -25,7 +25,6 @@ from wsidicom import WsiDicom
 
 
 @pytest.mark.integration
-@pytest.mark.filterwarnings("ignore")
 @parameterized_class(
     [
         {"input_type": WsiInputType.FILE},