+++ wsidicom/instance/image_coordinate_system.py
@@ -12,7 +12,6 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
-import warnings
 from typing import List, Optional, TypeVar
 
 from pydicom.dataset import Dataset