+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Use logger instead of issuing warnings.
+
 ### Fixed
 
 - Error in documentation of read_overview() in readme.