+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.0] - 2023-08-09
+
 ### Added
 
 - Override image size for labels and overviews if it is likely to be wrong.
@@ -168,7 +170,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of wsidicom
 
-[Unreleased]: https://github.com/imi-bigpicture/wsidicom/compare/0.10.0..HEAD
+[Unreleased]: https://github.com/imi-bigpicture/wsidicom/compare/0.11.0..HEAD
+[0.11.0]: https://github.com/imi-bigpicture/wsidicom/compare/v0.10.0..v0.11.0
 [0.10.0]: https://github.com/imi-bigpicture/wsidicom/compare/v0.9.0..v0.10.0
 [0.9.0]: https://github.com/imi-bigpicture/wsidicom/compare/v0.8.0..v0.9.0
 [0.8.0]: https://github.com/imi-bigpicture/wsidicom/compare/v0.7.0..v0.8.0